// applets/sonarr/api/model/releaseResource.ts
/**
 * Generated by orval v6.11.1 🍺
 * Do not edit manually.
 * Sonarr
 * Sonarr API docs
 * OpenAPI spec version: 3.0.0
 */
import type { QualityModel } from "./qualityModel";
import type { Language } from "./language";
import type { ReleaseEpisodeResource } from "./releaseEpisodeResource";
import type { CustomFormatResource } from "./customFormatResource";
import type { AlternateTitleResource } from "./alternateTitleResource";
import type { DownloadProtocol } from "./downloadProtocol";

export interface ReleaseResource {
  id?: number;
  guid?: string | null;
  quality?: QualityModel;
  qualityWeight?: number;
  age?: number;
  ageHours?: number;
  ageMinutes?: number;
  size?: number;
  indexerId?: number;
  indexer?: string | null;
  releaseGroup?: string | null;
  subGroup?: string | null;
  releaseHash?: string | null;
  title?: string | null;
  fullSeason?: boolean;
  sceneSource?: boolean;
  seasonNumber?: number;
  languages?: Language[] | null;
  languageWeight?: number;
  airDate?: string | null;
  seriesTitle?: string | null;
  episodeNumbers?: number[] | null;
  absoluteEpisodeNumbers?: number[] | null;
  mappedSeasonNumber?: number | null;
  mappedEpisodeNumbers?: number[] | null;
  mappedAbsoluteEpisodeNumbers?: number[] | null;
  mappedSeriesId?: number | null;
  mappedEpisodeInfo?: ReleaseEpisodeResource[] | null;
  approved?: boolean;
  temporarilyRejected?: boolean;
  rejected?: boolean;
  tvdbId?: number;
  tvRageId?: number;
  rejections?: string[] | null;
  publishDate?: string;
  commentUrl?: string | null;
  downloadUrl?: string | null;
  infoUrl?: string | null;
  episodeRequested?: boolean;
  downloadAllowed?: boolean;
  releaseWeight?: number;
  customFormats?: CustomFormatResource[] | null;
  customFormatScore?: number;
  sceneMapping?: AlternateTitleResource;
  magnetUrl?: string | null;
  infoHash?: string | null;
  seeders?: number | null;
  leechers?: number | null;
  protocol?: DownloadProtocol;
  isDaily?: boolean;
  isAbsoluteNumbering?: boolean;
  isPossibleSpecialEpisode?: boolean;
  special?: boolean;
  seriesId?: number | null;
  episodeId?: number | null;
  episodeIds?: number[] | null;
  downloadClientId?: number | null;
  shouldOverride?: boolean | null;
}